+++ .gitignore
@@ -35,3 +35,4 @@ node_modules/
 
 # WP-CLI
 vendor/
+.npmrc
